edit:テンプレートフォルダの制定
