+++ web/template/components/MyImg.tsx
@@ -1,0 +1,18 @@
+import Image from "next/image";
+import { FC, CSSProperties } from "react";
+
+interface MyImgProps {
+  containerStyles: string;
+  imgSrc: string;
+  style?: CSSProperties;
+}
+
+const MyImg: FC<MyImgProps> = ({ containerStyles, imgSrc, style }) => {
+  return (
+    <div className={`${containerStyles}`} style={style}>
+      <Image src={imgSrc} fill priority alt='' />
+    </div>
+  );
+};
+
+export default MyImg
